requirements.txt
Flask
Gunicorn
Skpy
flask-cors
python-dotenv 
PyJWT